+++ cms/src/migrations/index.ts
@@ -1,0 +1,9 @@
+import * as migration_20260918_065458_iniziale from './20260918_065458_iniziale';
+
+export const migrations = [
+  {
+    up: migration_20260918_065458_iniziale.up,
+    down: migration_20260918_065458_iniziale.down,
+    name: '20260918_065458_iniziale'
+  },
+];